♻️ refactor: Rename prop files and update StatisticName
- In this commit, I have renamed 'group.ts' to 'groupProps.ts' and 'member.ts' to 'memberProps.ts'.
- Additionally, I've updated 'PostName' to 'StatisticName' in statisticCardProps.ts for better clarity and consistency.

Related issue: #30
