+++ src/interfaces/group/index.ts
@@ -2,8 +2,8 @@ import CardProps from "./cardProps";
 import MemberCardProps from "./memberCardProps";
 import ScheduleCardProps from "./scheduleCardProps";
 import StatisticCardProps from "./statisticCardProps";
-import MemberProps from "./member";
-import GroupProps from "./group";
+import MemberProps from "./memberProps";
+import GroupProps from "./groupProps";
 
 export type {
   CardProps,
