+++ src/interfaces/group/memberProps.ts
